# docs/contracts/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Educational Framework Backend API",
    "description": "Family module API contract.",
    "version": "0.0.1"
  },
...
  "paths": {
...
    "/api/v1/dev/content/categories": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create category",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCategoryRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseCategory" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "409": { "description": "Conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error409" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List all categories",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseCategoryList" } } } }
        }
      }
    },
    "/api/v1/dev/content/categories/{id}": {
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Get category by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseCategory" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      },
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update category",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCategoryRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseCategory" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
    "/api/v1/dev/content/activities": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create activity",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateActivityRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivity" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "409": { "description": "Conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error409" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List activities, optionally filtered by topic",
        "parameters": [{ "name": "topicId", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivityList" } } } }
        }
      }
    },
    "/api/v1/dev/content/activities/{id}": {
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Get activity by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivity" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      },
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update activity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateActivityRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivity" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
    "/api/v1/dev/content/activity-resources": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create activity resource",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateActivityResourceRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivityResource" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found (activity not found)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List resources for an activity",
        "parameters": [{ "name": "activityId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivityResourceList" } } } }
        }
      }
    },
    "/api/v1/dev/content/activity-resources/{id}": {
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update activity resource",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateActivityResourceRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseActivityResource" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque family session token returned only by POST /api/v1/auth/login."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "data": { "nullable": true },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "displayOrder": { "type": "integer", "nullable": true },
          "iconUrl": { "type": "string", "nullable": true }
        }
      },
      "UpdateCategoryRequest": {
        "type": "object",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "displayOrder": { "type": "integer", "nullable": true },
          "iconUrl": { "type": "string", "nullable": true }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "displayOrder": { "type": "integer", "nullable": true },
          "iconUrl": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseCategory": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/CategoryResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseCategoryList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/CategoryResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateActivityRequest": {
        "type": "object",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "gameEngineType": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "topicIds": { "type": "array", "nullable": true, "items": { "type": "integer", "format": "int64" } }
        }
      },
      "UpdateActivityRequest": {
        "type": "object",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "gameEngineType": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "topicIds": { "type": "array", "nullable": true, "items": { "type": "integer", "format": "int64" } }
        }
      },
      "ActivityResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "gameEngineType": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "topicIds": { "type": "array", "items": { "type": "integer", "format": "int64" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseActivity": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ActivityResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseActivityList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateActivityResourceRequest": {
        "type": "object",
        "required": ["activityId", "resourceType", "path"],
        "properties": {
          "activityId": { "type": "integer", "format": "int64" },
          "topicId": { "type": "integer", "format": "int64", "nullable": true },
          "resourceType": { "type": "string", "enum": ["IMAGE", "AUDIO", "VIDEO"] },
          "path": { "type": "string", "maxLength": 500 },
          "metadata": { "type": "string", "nullable": true, "description": "JSON string with size, format, duration, dimensions, alt text." }
        }
      },
      "UpdateActivityResourceRequest": {
        "type": "object",
        "required": ["resourceType", "path"],
        "properties": {
          "topicId": { "type": "integer", "format": "int64", "nullable": true },
          "resourceType": { "type": "string", "enum": ["IMAGE", "AUDIO", "VIDEO"] },
          "path": { "type": "string", "maxLength": 500 },
          "metadata": { "type": "string", "nullable": true }
        }
      },
      "ActivityResourceResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "activityId": { "type": "integer", "format": "int64" },
          "topicId": { "type": "integer", "format": "int64", "nullable": true },
          "resourceType": { "type": "string", "enum": ["IMAGE", "AUDIO", "VIDEO"] },
          "path": { "type": "string" },
          "metadata": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseActivityResource": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ActivityResourceResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseActivityResourceList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityResourceResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "Error400": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Bad Request error response."
      },
...
      "Error404": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Not Found error response."
      },
      "Error409": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Conflict error response."
      },
...
    }
  }
}