# framework/backend/src/main/java/es/vargontoc/educational/framework/content/infrastructure/dto/UpdateActivityRequest.java
package es.vargontoc.educational.framework.content.infrastructure.dto;

import java.util.List;

public record UpdateActivityRequest(
    String name,
    String description,
    String gameEngineType,
    String status,
    Integer minAge,
    Integer maxAge,
    List<Long> topicIds
) {}